+++ test/xml_svg_test.dart
@@ -1,11 +1,24 @@
 import 'dart:ui';
 
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:flutter_svg/src/svg/xml_parsers.dart';
+import 'package:flutter_svg/src/svg/parser_state.dart';
 import 'package:flutter_svg/src/utilities/xml.dart';
 import 'package:path_drawing/path_drawing.dart';
 import 'package:test/test.dart';
 import 'package:xml/xml_events.dart';
+
+class TestSvgParserState extends SvgParserState {
+  TestSvgParserState({double fontSize = 14, double? xHeight})
+      : super(
+          <XmlEvent>[],
+          SvgTheme(fontSize: fontSize, xHeight: xHeight),
+          'testKey',
+          false,
+        );
+
+  @override
+  late Map<String, String> attributes;
+}
 
 void main() {
   test('Xlink href tests', () {
@@ -80,77 +93,39 @@ void main() {
     final XmlStartElementEvent svgWithNoSizeInfo =
         parseEvents('<svg />').first as XmlStartElementEvent;
 
+    final TestSvgParserState parserState = TestSvgParserState();
+
+    parserState.attributes =
+        svgWithViewBoxAndWidthHeight.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.size, const Size(50, 50));
+
+    parserState.attributes = svgWithViewBox.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.viewBoxRect, rect);
+
+    parserState.attributes = svgWithViewBox.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.viewBoxOffset, Offset.zero);
+
+    parserState.attributes =
+        svgWithViewBoxAndWidthHeight.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.viewBoxRect, rect);
+
+    parserState.attributes = svgWithWidthHeight.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.viewBoxRect, rect);
+
+    parserState.attributes = svgWithNoSizeInfo.attributes.toAttributeMap();
+    expect(parserState.parseViewBox(nullOk: true), null);
+
+    parserState.attributes = svgWithNoSizeInfo.attributes.toAttributeMap();
+    expect(() => parserState.parseViewBox(), throwsStateError);
+
+    parserState.attributes = svgWithViewBoxMinXMinY.attributes.toAttributeMap();
+    expect(parserState.parseViewBox()!.viewBoxRect, rect);
+
+    parserState.attributes = svgWithViewBoxMinXMinY.attributes.toAttributeMap();
     expect(
-        parseViewBox(
-          svgWithViewBoxAndWidthHeight.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .size,
-        const Size(50, 50));
-    expect(
-        parseViewBox(
-          svgWithViewBox.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxRect,
-        rect);
-    expect(
-        parseViewBox(
-          svgWithViewBox.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxOffset,
-        Offset.zero);
-    expect(
-        parseViewBox(
-          svgWithViewBoxAndWidthHeight.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxRect,
-        rect);
-    expect(
-        parseViewBox(
-          svgWithWidthHeight.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxRect,
-        rect);
-    expect(
-        parseViewBox(
-          svgWithNoSizeInfo.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-          nullOk: true,
-        ),
-        null);
-    expect(
-        () => parseViewBox(
-              svgWithNoSizeInfo.attributes.toAttributeMap(),
-              fontSize: 14.0,
-              xHeight: 7.0,
-            ),
-        throwsStateError);
-    expect(
-        parseViewBox(
-          svgWithViewBoxMinXMinY.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxRect,
-        rect);
-    expect(
-        parseViewBox(
-          svgWithViewBoxMinXMinY.attributes.toAttributeMap(),
-          fontSize: 14.0,
-          xHeight: 7.0,
-        )!
-            .viewBoxOffset,
-        const Offset(-42.0, -56.0));
+      parserState.parseViewBox()!.viewBoxOffset,
+      const Offset(-42.0, -56.0),
+    );
   });
 
   test('TileMode tests', () {
@@ -170,13 +145,21 @@ void main() {
     final XmlStartElementEvent none =
         parseEvents('<linearGradient />').first as XmlStartElementEvent;
 
-    expect(parseTileMode(pad.attributes.toAttributeMap()), TileMode.clamp);
-    expect(parseTileMode(invalid.attributes.toAttributeMap()), TileMode.clamp);
-    expect(parseTileMode(none.attributes.toAttributeMap()), TileMode.clamp);
+    final TestSvgParserState parserState = TestSvgParserState();
+    parserState.attributes = pad.attributes.toAttributeMap();
+    expect(parserState.parseTileMode(), TileMode.clamp);
 
-    expect(parseTileMode(reflect.attributes.toAttributeMap()), TileMode.mirror);
-    expect(
-        parseTileMode(repeat.attributes.toAttributeMap()), TileMode.repeated);
+    parserState.attributes = invalid.attributes.toAttributeMap();
+    expect(parserState.parseTileMode(), TileMode.clamp);
+
+    parserState.attributes = none.attributes.toAttributeMap();
+    expect(parserState.parseTileMode(), TileMode.clamp);
+
+    parserState.attributes = reflect.attributes.toAttributeMap();
+    expect(parserState.parseTileMode(), TileMode.mirror);
+
+    parserState.attributes = repeat.attributes.toAttributeMap();
+    expect(parserState.parseTileMode(), TileMode.repeated);
   });
 
   test('@stroke-dashoffset tests', () {
@@ -187,70 +170,80 @@ void main() {
         parseEvents('<stroke stroke-dashoffset="20%" />').first
             as XmlStartElementEvent;
 
+    final TestSvgParserState parserState = TestSvgParserState();
+    parserState.attributes = abs.attributes.toAttributeMap();
     expect(
-      parseDashOffset(
-        abs.attributes.toAttributeMap(),
-        fontSize: 14.0,
-        xHeight: 7.0,
-      ),
+      parserState.parseDashOffset(),
       equals(const DashOffset.absolute(20.0)),
     );
 
+    parserState.attributes = pct.attributes.toAttributeMap();
     expect(
-      parseDashOffset(
-        pct.attributes.toAttributeMap(),
-        fontSize: 14.0,
-        xHeight: 7.0,
-      ),
+      parserState.parseDashOffset(),
       equals(DashOffset.percentage(0.2)),
     );
   });
 
   test('font-weight tests', () {
-    expect(parseFontWeight('100'), FontWeight.w100);
-    expect(parseFontWeight('200'), FontWeight.w200);
-    expect(parseFontWeight('300'), FontWeight.w300);
-    expect(parseFontWeight('400'), FontWeight.w400);
-    expect(parseFontWeight('500'), FontWeight.w500);
-    expect(parseFontWeight('600'), FontWeight.w600);
-    expect(parseFontWeight('700'), FontWeight.w700);
-    expect(parseFontWeight('800'), FontWeight.w800);
-    expect(parseFontWeight('900'), FontWeight.w900);
+    final TestSvgParserState parserState = TestSvgParserState();
+    expect(parserState.parseFontWeight('100'), FontWeight.w100);
+    expect(parserState.parseFontWeight('200'), FontWeight.w200);
+    expect(parserState.parseFontWeight('300'), FontWeight.w300);
+    expect(parserState.parseFontWeight('400'), FontWeight.w400);
+    expect(parserState.parseFontWeight('500'), FontWeight.w500);
+    expect(parserState.parseFontWeight('600'), FontWeight.w600);
+    expect(parserState.parseFontWeight('700'), FontWeight.w700);
+    expect(parserState.parseFontWeight('800'), FontWeight.w800);
+    expect(parserState.parseFontWeight('900'), FontWeight.w900);
 
-    expect(parseFontWeight('normal'), FontWeight.normal);
-    expect(parseFontWeight('bold'), FontWeight.bold);
+    expect(parserState.parseFontWeight('normal'), FontWeight.normal);
+    expect(parserState.parseFontWeight('bold'), FontWeight.bold);
 
-    expect(() => parseFontWeight('invalid'), throwsUnsupportedError);
+    expect(
+        () => parserState.parseFontWeight('invalid'), throwsUnsupportedError);
   });
 
   test('font-style tests', () {
-    expect(parseFontStyle('normal'), FontStyle.normal);
-    expect(parseFontStyle('italic'), FontStyle.italic);
-    expect(parseFontStyle('oblique'), FontStyle.italic);
+    final TestSvgParserState parserState = TestSvgParserState();
+    expect(parserState.parseFontStyle('normal'), FontStyle.normal);
+    expect(parserState.parseFontStyle('italic'), FontStyle.italic);
+    expect(parserState.parseFontStyle('oblique'), FontStyle.italic);
 
-    expect(parseFontStyle(null), isNull);
-    expect(() => parseFontStyle('invalid'), throwsUnsupportedError);
+    expect(parserState.parseFontStyle(null), isNull);
+    expect(() => parserState.parseFontStyle('invalid'), throwsUnsupportedError);
   });
 
   test('text-decoration tests', () {
-    expect(parseTextDecoration('none'), TextDecoration.none);
-    expect(parseTextDecoration('line-through'), TextDecoration.lineThrough);
-    expect(parseTextDecoration('overline'), TextDecoration.overline);
-    expect(parseTextDecoration('underline'), TextDecoration.underline);
+    final TestSvgParserState parserState = TestSvgParserState();
+    expect(parserState.parseTextDecoration('none'), TextDecoration.none);
+    expect(parserState.parseTextDecoration('line-through'),
+        TextDecoration.lineThrough);
+    expect(
+        parserState.parseTextDecoration('overline'), TextDecoration.overline);
+    expect(
+        parserState.parseTextDecoration('underline'), TextDecoration.underline);
 
-    expect(parseTextDecoration(null), isNull);
-    expect(() => parseTextDecoration('invalid'), throwsUnsupportedError);
+    expect(parserState.parseTextDecoration(null), isNull);
+    expect(() => parserState.parseTextDecoration('invalid'),
+        throwsUnsupportedError);
   });
 
   test('text-decoration-style tests', () {
-    expect(parseTextDecorationStyle('solid'), TextDecorationStyle.solid);
-    expect(parseTextDecorationStyle('dashed'), TextDecorationStyle.dashed);
-    expect(parseTextDecorationStyle('dotted'), TextDecorationStyle.dotted);
-    expect(parseTextDecorationStyle('double'), TextDecorationStyle.double);
-    expect(parseTextDecorationStyle('wavy'), TextDecorationStyle.wavy);
+    final TestSvgParserState parserState = TestSvgParserState();
+    expect(parserState.parseTextDecorationStyle('solid'),
+        TextDecorationStyle.solid);
+    expect(parserState.parseTextDecorationStyle('dashed'),
+        TextDecorationStyle.dashed);
+    expect(parserState.parseTextDecorationStyle('dotted'),
+        TextDecorationStyle.dotted);
+    expect(parserState.parseTextDecorationStyle('double'),
+        TextDecorationStyle.double);
+    expect(
+        parserState.parseTextDecorationStyle('wavy'), TextDecorationStyle.wavy);
 
-    expect(parseTextDecorationStyle(null), isNull);
-    expect(() => parseTextDecorationStyle('invalid'), throwsUnsupportedError);
+    expect(parserState.parseTextDecorationStyle(null), isNull);
+    expect(() => parserState.parseTextDecorationStyle('invalid'),
+        throwsUnsupportedError);
   });
 
   group('parseStyle', () {
@@ -260,15 +253,11 @@ void main() {
           parseEvents('<svg stroke="currentColor" />').first
               as XmlStartElementEvent;
 
-      final DrawableStyle svgStyle = parseStyle(
-        'test',
-        svg.attributes.toAttributeMap(),
-        DrawableDefinitionServer(),
-        null,
+      final TestSvgParserState parserState = TestSvgParserState();
+      parserState.attributes = svg.attributes.toAttributeMap();
+      final DrawableStyle svgStyle = parserState.parseStyle(
         null,
         currentColor: currentColor,
-        fontSize: 14.0,
-        xHeight: 7.0,
       );
 
       expect(
@@ -283,15 +272,11 @@ void main() {
           parseEvents('<svg fill="currentColor" />').first
               as XmlStartElementEvent;
 
-      final DrawableStyle svgStyle = parseStyle(
-        'test',
-        svg.attributes.toAttributeMap(),
-        DrawableDefinitionServer(),
-        null,
+      final TestSvgParserState parserState = TestSvgParserState();
+      parserState.attributes = svg.attributes.toAttributeMap();
+      final DrawableStyle svgStyle = parserState.parseStyle(
         null,
         currentColor: currentColor,
-        fontSize: 14.0,
-        xHeight: 7.0,
       );
 
       expect(
@@ -308,15 +293,11 @@ void main() {
 
         const double fontSize = 26.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
-          xHeight: 7.0,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           svgStyle.stroke?.strokeWidth,
@@ -331,15 +312,11 @@ void main() {
 
         const double fontSize = 26.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
-          xHeight: 7.0,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           <double>[
@@ -362,15 +339,11 @@ void main() {
 
         const double fontSize = 26.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
-          xHeight: 7.0,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           svgStyle.dashOffset,
@@ -388,15 +361,12 @@ void main() {
         const double fontSize = 26.0;
         const double xHeight = 11.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
           xHeight: xHeight,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           svgStyle.stroke?.strokeWidth,
@@ -412,15 +382,12 @@ void main() {
         const double fontSize = 26.0;
         const double xHeight = 11.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
           xHeight: xHeight,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           <double>[
@@ -444,15 +411,12 @@ void main() {
         const double fontSize = 26.0;
         const double xHeight = 11.0;
 
-        final DrawableStyle svgStyle = parseStyle(
-          'test',
-          svg.attributes.toAttributeMap(),
-          DrawableDefinitionServer(),
-          null,
-          null,
+        final TestSvgParserState parserState = TestSvgParserState(
           fontSize: fontSize,
           xHeight: xHeight,
         );
+        parserState.attributes = svg.attributes.toAttributeMap();
+        final DrawableStyle svgStyle = parserState.parseStyle(null);
 
         expect(
           svgStyle.dashOffset,
